+++ metrics/collectors/shared.lua
@@ -2,7 +2,70 @@ local clock = require('clock')
 local fiber = require('fiber')
 local log = require('log')
 
-local Shared = {}
+local Shared = {Prepared = {}}
+
+function Shared.Prepared:new_class(method_names)
+    local methods = {}
+    for _, name in ipairs(method_names or {}) do
+        methods[name] = Shared.Prepared[name]
+    end
+    local class = {}
+    class.__index = class
+    return setmetatable(class, {__index = methods})
+end
+
+function Shared.Prepared:new(collector, label_pairs)
+    return setmetatable({
+        collector = collector,
+        label_pairs = label_pairs,
+        -- `make_key` is a pretty heavy method since it works with strings intensively
+        -- The idea is to cache the key and re-use it for all the "prepared" statements
+        key = collector:make_key(label_pairs),
+    }, self)
+end
+
+function Shared.Prepared:remove()
+    assert(self.label_pairs, 'label pairs is a required parameter')
+    self.collector.observations[self.key] = nil
+    self.collector.label_pairs[self.key] = nil
+end
+
+function Shared.Prepared:set(num)
+    if num ~= nil and type(tonumber(num)) ~= 'number' then
+        error("Collector set value should be a number")
+    end
+    num = num or 0
+    self.collector.observations[self.key] = num
+    self.collector.label_pairs[self.key] = self.label_pairs or {}
+end
+
+function Shared.Prepared:inc(num)
+    if num ~= nil and type(tonumber(num)) ~= 'number' then
+        error("Collector increment should be a number")
+    end
+    num = num or 1
+    local old_value = self.collector.observations[self.key] or 0
+    self.collector.observations[self.key] = old_value + num
+    self.collector.label_pairs[self.key] = self.label_pairs or {}
+end
+
+function Shared.Prepared:dec(num)
+    if num ~= nil and type(tonumber(num)) ~= 'number' then
+        error("Collector decrement should be a number")
+    end
+    num = num or 1
+    local old_value = self.collector.observations[self.key] or 0
+    self.collector.observations[self.key] = old_value - num
+    self.collector.label_pairs[self.key] = self.label_pairs or {}
+end
+
+function Shared.Prepared:observe()
+    error('Not implemented in shared class, override me')
+end
+
+function Shared.Prepared:reset()
+    error('Not implemented in shared class, override me')
+end
 
 -- Create collector class with the list of instance methods copied from
 -- this class (like an inheritance but with limited list of methods).
@@ -12,6 +75,7 @@ function Shared:new_class(kind, method_names)
     table.insert(method_names, 'new')
     table.insert(method_names, 'set_registry')
     table.insert(method_names, 'make_key')
+    table.insert(method_names, 'prepare')
     table.insert(method_names, 'append_global_labels')
     table.insert(method_names, 'collect')
     table.insert(method_names, 'remove')
@@ -19,7 +83,10 @@ function Shared:new_class(kind, method_names)
     for _, name in pairs(method_names) do
         methods[name] = self[name]
     end
-    local class = {kind = kind}
+    local class = {
+        kind = kind,
+        Prepared = Shared.Prepared:new_class(method_names),
+    }
     class.__index = class
     return setmetatable(class, {__index = methods})
 end
@@ -85,43 +152,32 @@ function Shared:make_key(label_pairs)
     return table.concat(parts, '\t')
 end
 
+function Shared:prepare(label_pairs)
+    return self.Prepared:new(self, label_pairs)
+end
+
 function Shared:remove(label_pairs)
-    assert(label_pairs, 'label pairs is a required parameter')
-    local key = self:make_key(label_pairs)
-    self.observations[key] = nil
-    self.label_pairs[key] = nil
+    self:prepare(label_pairs):remove()
 end
 
 function Shared:set(num, label_pairs)
-    if num ~= nil and type(tonumber(num)) ~= 'number' then
-        error("Collector set value should be a number")
-    end
-    num = num or 0
-    local key = self:make_key(label_pairs)
-    self.observations[key] = num
-    self.label_pairs[key] = label_pairs or {}
+    self:prepare(label_pairs):set(num)
 end
 
 function Shared:inc(num, label_pairs)
-    if num ~= nil and type(tonumber(num)) ~= 'number' then
-        error("Collector increment should be a number")
-    end
-    num = num or 1
-    local key = self:make_key(label_pairs)
-    local old_value = self.observations[key] or 0
-    self.observations[key] = old_value + num
-    self.label_pairs[key] = label_pairs or {}
+    self:prepare(label_pairs):inc(num)
 end
 
 function Shared:dec(num, label_pairs)
-    if num ~= nil and type(tonumber(num)) ~= 'number' then
-        error("Collector decrement should be a number")
-    end
-    num = num or 1
-    local key = self:make_key(label_pairs)
-    local old_value = self.observations[key] or 0
-    self.observations[key] = old_value - num
-    self.label_pairs[key] = label_pairs or {}
+    self:prepare(label_pairs):dec(num)
+end
+
+function Shared:observe(num, label_pairs)
+    self:prepare(label_pairs):observe(num)
+end
+
+function Shared:reset(label_pairs)
+    self:prepare(label_pairs):reset()
 end
 
 local function log_observe_latency_error(err)